Warning fix: Instead of casting long to int and then using Integer.toString, just use Long.toString.

+++ src/controller/map/drivers/QueryCLI.java
@@ -154,7 +154,7 @@ public final class QueryCLI implements ISPDriver {
 			ydiff = dims.cols - rawYdiff;
 		}
 		ostream.append("Distance (as the crow flies, in tiles):\t");
-		ostream.append(Integer.toString((int) Math.round(Math.sqrt(xdiff
+		ostream.append(Long.toString(Math.round(Math.sqrt(xdiff
 				* xdiff + ydiff * ydiff))));
 	}
 	/**
